Rename: tiket -> ticket 폴더 및 이름 오타 수정
